# openapi.yaml
openapi: 3.1.0
info:
  title: SecureFlow API
  description: |
    SecureFlow is a GitHub App that provides automated security scanning for pull requests.
    This API enables programmatic access to findings, repositories, SBOM scanning, and more.
  version: 1.0.0
  contact:
    name: SecureFlow Team
    url: https://github.com/GauravKarakoti/SecureFlow
  license:
    name: MIT
    url: https://github.com/GauravKarakoti/SecureFlow/blob/main/LICENSE
servers:
  - url: https://secure-flow-six.vercel.app
    description: Production server
  - url: http://localhost:9002
    description: Local development server
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: |
        NextAuth.js JWT session cookie. Obtained via GitHub OAuth flow.
        Sessions have a 1-year maxAge. The session contains user ID, codename, and roles.
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Machine-readable error code
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - UNPROCESSABLE_ENTITY
            - TOO_MANY_REQUESTS
            - INTERNAL_SERVER_ERROR
            - DATABASE_ERROR
        message:
          type: string
          description: Human-readable error message (sensitive data scrubbed)
        success:
          type: boolean
          enum: [false]
    SbomDependency:
      type: object
      required:
        - name
        - version
        - manifestFile
        - ecosystem
      properties:
        name:
          type: string
          description: Package name
          example: express
        version:
          type: string
          description: Package version
          example: 4.18.0
        manifestFile:
          type: string
          description: Source manifest file
          example: package.json
        ecosystem:
          type: string
          enum: [npm, pypi, maven, gem]
          description: Package ecosystem
    VulnerabilityMatch:
      type: object
      required:
        - dependency
        - cveId
        - severity
        - description
      properties:
        dependency:
          $ref: "#/components/schemas/SbomDependency"
        cveId:
          type: string
          description: CVE identifier
          example: CVE-2021-44228
        severity:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        description:
          type: string
          description: Vulnerability description
        patchedVersion:
          type: string
          nullable: true
          description: Version that fixes the vulnerability
    SbomScanResult:
      type: object
      required:
        - scanId
        - timestamp
        - totalDependencies
        - vulnerabilities
        - status
      properties:
        scanId:
          type: string
          format: uuid
          description: Unique scan identifier
        timestamp:
          type: string
          format: date-time
          description: ISO-8601 UTC timestamp
        totalDependencies:
          type: integer
          minimum: 0
          description: Total number of dependencies found
        vulnerabilities:
          type: array
          items:
            $ref: "#/components/schemas/VulnerabilityMatch"
        status:
          type: string
          enum: [CLEAN, WARNING, VULNERABLE, ERROR]
          description: Overall scan status
...
    ScanJobRequest:
      type: object
      required:
        - repositoryId
      properties:
        repositoryId:
          type: string
          description: SecureFlow repository ID (CUID)
        prNumber:
          type: integer
          description: Pull request number (optional)
...
paths:
  /api/sbom/scan:
    post:
      summary: Scan SBOM for vulnerabilities
      description: |
        Analyzes a manifest file (package.json, requirements.txt, etc.) for known vulnerabilities
        in dependencies. Returns a list of vulnerable packages with CVE details.
      operationId: scanSbom
      security:
        - sessionAuth: []
      tags:
        - SBOM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fileName
                - content
              properties:
                fileName:
                  type: string
                  description: Name of the manifest file
                  example: package.json
                content:
                  type: string
                  description: Raw content of the manifest file
            examples:
              npm:
                summary: NPM package.json
                value:
                  fileName: package.json
                  content: |
                    {
                      "dependencies": {
                        "express": "4.17.1",
                        "lodash": "4.17.19"
                      }
                    }
      responses:
        "200":
          description: Scan completed successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - result
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  result:
                    $ref: "#/components/schemas/SbomScanResult"
        "400":
          description: Invalid request - missing fileName or content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized - session required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/findings:
    post:
      summary: Enqueue vulnerability scan
      description: |
        Enqueues an asynchronous vulnerability scan for a repository.
        Returns immediately with a job ID for status polling.
      operationId: enqueueScan
      security:
        - sessionAuth: []
      tags:
        - Findings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanJobRequest"
      responses:
        "202":
          description: Scan job enqueued
          headers:
            Cache-Control:
              schema:
                type: string
                default: no-store
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [queued]
                  jobId:
                    type: string
                    description: BullMQ job ID
                  scanJobId:
                    type: string
                    description: Scan job ID for status polling
                  message:
                    type: string
                  pollingUrl:
                    type: string
                    description: URL to poll for status
                    example: /api/findings/status/job_123
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Repository not found or not owned by user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded (10 req/60s)
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...